Add API Get Nodes check
Use a OCP Service Account to generate a token for accessing the API.
This token is then stored in a vault file for use by the playbook to
create a file on the host for use in the CURL command. The creation of
the Service Account is documented in
https://github.houston.entsvcs.net/steven-barre/openshift-docs/blob/master/API-Monitor.md

The .vaultpass file is to be ignored by git and should never be comitted
to git.

Added a README with directions for deploying the monitor and where to
get the vault password.